886 part4

+++ src/rockstor/smart_manager/replication/receiver.py
@@ -19,10 +19,8 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 from multiprocessing import Process
 import os
 import sys
-import logging
 import zmq
 import subprocess
-from datetime import datetime
 from django.utils.timezone import utc
 from django.conf import settings
 from contextlib import contextmanager
@@ -32,15 +30,15 @@ from system.osi import run_command
 from storageadmin.models import (Disk, Pool, Appliance)
 from smart_manager.models import ReplicaShare
 import shutil
-
+import time
+from cli import APIWrapper
 
 BTRFS = '/sbin/btrfs'
-logger = logging.getLogger(__name__)
 
 
 class Receiver(ReplicationMixin, Process):
 
-    def __init__(self, meta):
+    def __init__(self, meta, logger):
         self.meta = meta
         self.meta_port = self.meta['meta_port']
         self.data_port = self.meta['data_port']
@@ -57,9 +55,18 @@ class Receiver(ReplicationMixin, Process):
         self.meta_push = None
         self.num_retain_snaps = 5
         self.ctx = zmq.Context()
+        self.logger = logger
+        self.rp = None
+        self.raw = None
         super(Receiver, self).__init__()
 
+
     def _sys_exit(self, code, linger=60000):
+        if (self.rp is not None and self.rp.returncode is None):
+            try:
+                self.rp.terminate()
+            except Exception, e:
+                self.logger.error('Exception while terminating the btrfs-recv process: %s' % e.__str__())
         self.ctx.destroy(linger=linger)
         sys.exit(code)
 
@@ -68,17 +75,18 @@ class Receiver(ReplicationMixin, Process):
         try:
             yield
         except Exception, e:
-            logger.error('%s. Exception: %s' % (msg, e.__str__()))
+            self.logger.error('%s. Exception: %s' % (msg, e.__str__()))
             if (ack is True):
                 try:
                     err_ack = {'msg': 'error',
                                'id': self.meta['id'],
                                'error': msg, }
                     self.meta_push.send_json(err_ack)
+                    self.logger.debug('error_ack sent: %s' % err_ack)
                 except Exception, e:
                     msg = ('Failed to send ack: %s to the sender for meta: '
                            '%s. Aborting' % (err_ack, self.meta))
-                    logger.error('%s. Exception: %s' % (msg, e.__str__()))
+                    self.logger.error('%s. Exception: %s' % (msg, e.__str__()))
                     self._sys_exit(3)
             self._sys_exit(3)
 
@@ -88,11 +96,12 @@ class Receiver(ReplicationMixin, Process):
             msg = ('Failed to delete snapshot: %s. Aborting.' %
                    oldest_snap)
             with self._clean_exit_handler(msg):
-                self.delete_snapshot(share_name, oldest_snap)
-                return self._delete_old_snaps(share_name, share_path, num_retain)
+                if (self.delete_snapshot(share_name, oldest_snap, self.logger)):
+                    return self._delete_old_snaps(share_name, share_path, num_retain)
 
 
     def run(self):
+        self.law = APIWrapper()
         msg = ('Failed to get the sender ip from the uuid(%s) for meta: %s' %
                (self.meta['uuid'], self.meta))
         with self._clean_exit_handler(msg):
@@ -113,8 +122,8 @@ class Receiver(ReplicationMixin, Process):
             #in the db, no error is raised here.
             recv_sub = self.ctx.socket(zmq.SUB)
             recv_sub.connect('tcp://%s:%d' % (self.sender_ip, self.data_port))
-            recv_sub.RCVTIMEO = 1000 # 1 second
-            recv_sub.setsockopt(zmq.SUBSCRIBE, str(self.meta['id']))
+            recv_sub.RCVTIMEO = 1000 # 1 minute
+            recv_sub.setsockopt_string(zmq.SUBSCRIBE, self.meta['id'].decode('ascii'))
 
         msg = ('Failed to connect to the sender(%s) on '
                'meta_port(%d). meta: %s. Aborting.' %
@@ -130,7 +139,7 @@ class Receiver(ReplicationMixin, Process):
             msg = ('Failed to verify/create share: %s. meta: %s. '
                    'Aborting.' % (sname, self.meta))
             with self._clean_exit_handler(msg, ack=True):
-                self.create_share(sname, self.dest_pool)
+                self.create_share(sname, self.dest_pool, self.logger)
 
             msg = ('Failed to create the replica metadata object '
                    'for share: %s. meta: %s. Aborting.' %
@@ -165,14 +174,16 @@ class Receiver(ReplicationMixin, Process):
                 ack = {'msg': 'snap_exists',
                        'id': self.meta['id'], }
                 self.meta_push.send_json(ack)
+                self.logger.debug('snap_exists ack sent: %s' % ack)
+                self._sys_exit(0)
 
         cmd = [BTRFS, 'receive', snap_dir]
         msg = ('Failed to start the low level btrfs receive command(%s)'
                '. Aborting.' % (cmd))
         with self._clean_exit_handler(msg, ack=True):
-            rp = subprocess.Popen(cmd, shell=False, stdin=subprocess.PIPE,
-                                  stdout=subprocess.PIPE,
-                                  stderr=subprocess.PIPE)
+            self.rp = subprocess.Popen(cmd, shell=False, stdin=subprocess.PIPE,
+                                       stdout=subprocess.PIPE,
+                                       stderr=subprocess.PIPE)
 
         msg = ('Failed to send begin_ok to the sender for meta: %s' %
                self.meta)
@@ -180,32 +191,52 @@ class Receiver(ReplicationMixin, Process):
             ack = {'msg': 'begin_ok',
                    'id': self.meta['id'], }
             self.meta_push.send_json(ack)
+            self.logger.debug('begin_ok ack sent: %s' % ack)
         recv_timeout_counter = 0
         credit = settings.DEFAULT_SEND_CREDIT
         check_credit = True
+        t0 = t0_cycle = time.time()
+        recv_cycle = 0
+        msg_count = 0
         while True:
             if (check_credit is True and credit < 5):
                 ack = {'msg': 'send_more',
                        'id': self.meta['id'],
                        'credit': settings.DEFAULT_SEND_CREDIT, }
                 self.meta_push.send_json(ack)
+                self.logger.debug('send_more ack sent: %s' % ack)
+                msg_count_cycle = settings.DEFAULT_SEND_CREDIT - credit
+                msg_count += msg_count_cycle
+                cur_t = time.time()
+                total_kb = self.kb_received / 1024
+                cycle_kb = recv_cycle / 1024
+                total_xfer = total_kb / (cur_t - t0)
+                cycle_xfer = cycle_kb / (cur_t - t0_cycle)
+                total_avg_msg = self.kb_received / msg_count
+                cycle_avg_msg = recv_cycle / msg_count_cycle
+                self.logger.debug('%s KB received so far: %f . this cycle: %f .'
+                                  'xfer rate so far: %f . this cycle: %f .'
+                                  'Avg msg size(Bytes) so far: %f . this cycle: %f .' %
+                                  (self.meta['id'], total_kb, cycle_kb, total_xfer, cycle_xfer,
+                                   total_avg_msg, cycle_avg_msg))
+                recv_cycle = 0
+                t0_cycle = time.time()
                 credit = credit + settings.DEFAULT_SEND_CREDIT
-                logger.debug('%d KB received for %s' %
-                             (int(self.kb_received / 1024), sname))
 
             try:
                 recv_data = recv_sub.recv()
                 recv_data = recv_data[len(self.meta['id']):]
                 credit = credit - 1
                 recv_timeout_counter = 0
-                self.kb_received = self.kb_received + len(recv_data)
+                self.kb_received += len(recv_data)
+                recv_cycle += len(recv_data)
                 if (self.rtid is None):
                     msg = ('Failed to create snapshot: %s. Aborting.' %
                            self.snap_name)
                     # create a snapshot only if it's not already from a previous failed attempt
                     with self._clean_exit_handler(msg, ack=True):
                         self.create_snapshot(sname, self.snap_name,
-                                             snap_type='receiver')
+                                             self.logger, snap_type='receiver')
 
                     data = {'snap_name': self.snap_name}
                     msg = ('Failed to create receive trail for rid: %d'
@@ -215,22 +246,28 @@ class Receiver(ReplicationMixin, Process):
 
                 if (recv_data == 'END_SUCCESS' or recv_data == 'END_FAIL'):
                     check_credit = False
-                    ts = datetime.utcnow().replace(tzinfo=utc)
                     data = {'kb_received': self.kb_received / 1024, }
+                    self.logger.debug('END message received for %s : %s' %
+                                      (self.meta['id'], recv_data))
                     if (recv_data == 'END_SUCCESS'):
-                        data['receive_succeeded'] = ts.strftime(settings.SNAP_TS_FORMAT)
-                        #delete any snapshots older than num_retain
-                        self._delete_old_snaps(sname, snap_dir, self.num_retain_snaps + 1)
+                        data['status'] = 'succeeded'
+                        try:
+                            #delete any snapshots older than num_retain
+                            self._delete_old_snaps(sname, snap_dir, self.num_retain_snaps + 1)
+                        except Exception, e:
+                            self.logger.error('Exception while deleting old '
+                                              'snapshots: %s' % e.__str__())
+                            #raising the exception would make a bigger mess.
+                            #problematic past snapshots can be manually deleted.
+
                         #delete the share, move the oldest snap to share
-                        oldest_snap = get_oldest_snap(snap_dir, self.num_retain_snaps)
-                        if (oldest_snap is not None):
-                            snap_path = ('%s/%s' % (snap_dir, oldest_snap))
-                            share_path = ('%s%s/%s' %
-                                          (settings.MNT_PT, self.dest_pool,
-                                           sname))
-                            msg = ('Failed to promote the oldest Snapshot(%s) '
-                                   'to Share(%s)' % (snap_path, share_path))
-                            try:
+                        try:
+                            oldest_snap = get_oldest_snap(snap_dir, self.num_retain_snaps)
+                            if (oldest_snap is not None):
+                                snap_path = ('%s/%s' % (snap_dir, oldest_snap))
+                                share_path = ('%s%s/%s' %
+                                              (settings.MNT_PT, self.dest_pool,
+                                               sname))
                                 pool = Pool.objects.get(name=self.dest_pool)
                                 remove_share(pool, sname, '-1/-1')
                                 set_property(snap_path, 'ro', 'false',
@@ -238,14 +275,16 @@ class Receiver(ReplicationMixin, Process):
                                 run_command(['/usr/bin/rm', '-rf', share_path],
                                             throw=False)
                                 shutil.move(snap_path, share_path)
-                                self.delete_snapshot(sname, oldest_snap)
-                            except Exception, e:
-                                logger.error('%s. Exception: %s' % (msg, e.__str__()))
+                                self.delete_snapshot(sname, oldest_snap, self.logger)
+                        except Exception, e:
+                            msg = ('Failed to promote the oldest Snapshot to Share'
+                                   ' for %s' % self.meta['id'])
+                            self.logger.error('%s. Exception: %s' % (msg, e.__str__()))
                     else:
-                        logger.error('END_FAIL received for meta: %s. '
-                                     'Terminating.' % self.meta)
-                        rp.terminate()
-                        data['receive_failed'] = ts.strftime(settings.SNAP_TS_FORMAT)
+                        self.logger.error('END_FAIL received for meta: %s. '
+                                          'Terminating.' % self.meta)
+                        if (self.rp.poll() is None):
+                            self.rp.terminate()
                         data['status'] = 'failed'
 
                     msg = ('Failed to update receive trail for rtid: %d'
@@ -253,38 +292,40 @@ class Receiver(ReplicationMixin, Process):
                     with self._clean_exit_handler(msg, ack=True):
                         self.update_receive_trail(self.rtid, data)
                     break
-                if (rp.poll() is None):
-                    rp.stdin.write(recv_data)
-                    rp.stdin.flush()
+                if (self.rp.poll() is None):
+                    self.rp.stdin.write(recv_data)
+                    self.rp.stdin.flush()
                 else:
-                    logger.error('It seems the btrfs receive process died'
-                                 ' unexpectedly.')
-                    out, err = rp.communicate()
+                    self.logger.error('It seems the btrfs receive process died'
+                                      ' unexpectedly for meta: %s' % self.meta)
+                    out, err = self.rp.communicate()
                     msg = ('Low level system error from btrfs receive '
                            'command. out: %s err: %s for rtid: %s meta: %s'
                            % (out, err, self.rtid, self.meta))
                     with self._clean_exit_handler(msg, ack=True):
-                        ts = datetime.utcnow().replace(tzinfo=utc)
-                        data = {'receive_failed': ts.strftime(settings.SNAP_TS_FORMAT),
-                                'status': 'failed',
+                        data = {'status': 'failed',
                                 'error': msg, }
                         self.update_receive_trail(self.rtid, data)
             except zmq.error.Again:
                 recv_timeout_counter = recv_timeout_counter + 1
                 if (recv_timeout_counter > 600): #10 minutes
-                    logger.error('Nothing received in the last 60 seconds '
-                                 'from the sender(%s) for meta: %s. Aborting.'
-                                 % (self.sender_ip, self.meta))
+                    self.logger.error('Nothing received in the last 60 seconds '
+                                      'from the sender(%s) for meta: %s. Aborting.'
+                                      % (self.sender_ip, self.meta))
                     self._sys_exit(3)
             except Exception, e:
-                msg = ('Exception occured while receiving fsdata: %s' % e.__str__())
-                logger.error(msg)
-                rp.terminate()
-                out, err = rp.communicate()
-                data['receive_failed'] = datetime.utcnow().replace(tzinfo=utc).strftime(settings.SNAP_TS_FORMAT)
+                msg = ('Exception occured while receiving fsdata for meta: %s.'
+                       'Exception: %s' % (self.meta, e.__str__()))
+                self.logger.error(msg)
+                try:
+                    self.rp.terminate()
+                except Exception, e:
+                    self.logger.error('Exception while terminating btrfs-recv '
+                                      'for %s: %s' % (self.meta['id'], e.__str__()))
+                    #don't raise the exception because it will create a bigger mess.
+
                 data['status'] = 'failed'
                 data['error'] = msg
-
                 msg = ('Failed to update receive trail for rtid: %d'
                        '. meta: %s' % (self.rtid, self.meta))
                 with self._clean_exit_handler(msg, ack=True):
@@ -292,22 +333,24 @@ class Receiver(ReplicationMixin, Process):
                 self._sys_exit(3)
             finally:
                 if (os.getppid() != self.ppid):
-                    logger.error('parent exited. aborting.')
+                    self.logger.error('parent exited. aborting.')
                     self._sys_exit(3)
 
         #rfo/stdin should be closed by now. We get here only if the sender
         #dint throw an error or if receiver did not get terminated
         try:
-            out, err = rp.communicate()
+            out, err = self.rp.communicate()
+            self.logger.debug('cmd = %s out: %s err: %s rc: %s' %
+                              (cmd, out, err, self.rp.returncode))
         except Exception, e:
-            logger.debug('Exception while terminating receive. Probably '
-                         'already terminated: %s' % e.__str__())
+            self.logger.debug('Exception while terminating receive. Meta: %s. '
+                              'Probably already terminated: %s' %
+                              (self.meta, e.__str__()))
 
         ack = {'msg': 'receive_ok',
                'id': self.meta['id'], }
-        data = {'status': 'succeeded',
-                'end_ts': datetime.utcnow().replace(tzinfo=utc).strftime(settings.SNAP_TS_FORMAT), }
-        if (rp.returncode != 0):
+        data = {'status': 'succeeded', }
+        if (self.rp.returncode != 0):
             ack['msg'] = 'receive_error'
             data['status'] = 'failed'
 
@@ -319,14 +362,18 @@ class Receiver(ReplicationMixin, Process):
         msg = ('Failed to send final ack to the sender for meta: %s' %
                self.meta)
         with self._clean_exit_handler(msg):
-            logger.debug('Receive finished for %s. ack = %s' % (sname, ack))
             self.meta_push.send_json(ack)
+            self.logger.debug('Receive finished for %s. ack = %s' % (sname, ack))
 
         try:
-            recv_sub.RCVTIMEO = 60000
+            recv_sub.RCVTIMEO = 60000 # 1 minute
             recv_data = recv_sub.recv()
             recv_data = recv_data[len(self.meta['id']):]
         except Exception, e:
-            logger.error('Exception while waiting for final ack from sender: %s' % e.__str__())
+            self.logger.error('Exception while waiting for final ack from '
+                              'sender for %s: %s' % (sname, e.__str__()))
+            #it's ok if we don't receive the final ack, this is perhaps a
+            #hacky way of waiting a while to make sure the previous send
+            #made it through.
 
         self._sys_exit(0)
