+++ src/rockstor/smart_manager/replication/sender.py
@@ -19,29 +19,27 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 from multiprocessing import Process
 import os
 import sys
-import logging
 import zmq
 import subprocess
 import fcntl
 from django.conf import settings
-from datetime import datetime
 from contextlib import contextmanager
 from django.utils.timezone import utc
 from util import ReplicationMixin
 from fs.btrfs import get_oldest_snap
 from storageadmin.models import Appliance
 import json
+from cli import APIWrapper
 
 BTRFS = '/sbin/btrfs'
-logger = logging.getLogger(__name__)
 
 
 class Sender(ReplicationMixin, Process):
 
-    def __init__(self, replica, sender_ip, snap_name, smeta_port,
-                 sdata_port, rmeta_port, uuid, snap_id, rt=None):
+    def __init__(self, replica, sender_ip, receiver_ip, snap_name, smeta_port,
+                 sdata_port, rmeta_port, uuid, snap_id, logger, rt=None):
         self.replica = replica
-        self.receiver_ip = self._get_receiver_ip(self.replica)
+        self.receiver_ip = receiver_ip
         self.smeta_port = smeta_port
         self.sdata_port = sdata_port
         self.rmeta_port = rmeta_port
@@ -68,27 +66,26 @@ class Sender(ReplicationMixin, Process):
                          'msg': 'end', }
         self.kb_sent = 0
         self.ctx = zmq.Context()
+        self.logger = logger
+        self.law = None
+        self.raw = None
+        self.sp = None
         super(Sender, self).__init__()
 
-    def _get_receiver_ip(self, replica):
-        try:
-            appliance = Appliance.objects.get(uuid=replica.appliance)
-            return appliance.ip
-        except Exception, e:
-            logger.exception(e)
-            msg = ('Failed to get receiver ip. Is the receiver '
-                   'appliance added?')
-            raise Exception(msg)
 
     @contextmanager
     def _clean_exit_handler(self, msg):
         try:
             yield
         except Exception, e:
-            logger.error('%s. Exception: %s' % (msg, e.__str__()))
+            self.logger.error('%s. Exception: %s' % (msg, e.__str__()))
             self._sys_exit(3)
 
     def _sys_exit(self, code, linger=60000):
+        if (self.sp is not None and
+            self.sp.poll() is None):
+            self.sp.terminate()
+
         self.ctx.destroy(linger=linger)
         sys.exit(code)
 
@@ -97,20 +94,20 @@ class Sender(ReplicationMixin, Process):
         try:
             yield
         except Exception, e:
-            logger.error('%s. Exception: %s' % (msg, e.__str__()))
+            self.logger.error('%s. Exception: %s' % (msg, e.__str__()))
             try:
                 data = {'status': 'failed',
-                        'error': msg,
-                        'end_ts': datetime.utcnow().replace(tzinfo=utc).strftime(settings.SNAP_TS_FORMAT), }
+                        'error': msg, }
                 self.update_replica_status(self.rt2_id, data)
             except Exception, e:
-                logger.error('Exception occured in cleanup handler: %s' % e.__str__())
+                self.logger.error('Exception occured in cleanup handler: %s' % e.__str__())
             finally:
                 self._sys_exit(3)
 
     def _process_q(self):
-        ack = self.meta_sub.recv()
+        ack = self.meta_sub.recv_string()
         ack = json.loads(ack[len('meta-%s' % self.snap_id):])
+        self.logger.debug('ack received by sender: %s' % ack)
         if (ack['msg'] == 'send_more'):
             #  excess credit messages from receiver at that end
             return self._process_q()
@@ -127,10 +124,20 @@ class Sender(ReplicationMixin, Process):
             msg = ('Failed to delete snapshot: %s. Aborting.' %
                    oldest_snap)
             with self._clean_exit_handler(msg):
-                self.delete_snapshot(self.replica.share, oldest_snap)
-                return self._delete_old_snaps(share_path)
+                if (self.delete_snapshot(self.replica.share, oldest_snap, self.logger)):
+                    return self._delete_old_snaps(share_path)
 
     def run(self):
+
+        msg = ('Failed to subscribe to the main scheduler')
+        with self._clean_exit_handler(msg):
+            self.law = APIWrapper()
+            self.meta_sub = self.ctx.socket(zmq.SUB)
+            self.meta_sub.connect('tcp://%s:%d' % (self.sender_ip, self.sdata_port))
+            self.meta_sub.RCVTIMEO = 60000 #1 minutes
+            substr = 'meta-%s' % self.snap_id
+            self.meta_sub.setsockopt_string(zmq.SUBSCRIBE, substr.decode('ascii'))
+
         msg = ('Failed to connect to receiver(%s) on meta port'
                '(%d) for snap_name: %s. Aborting.' %
                (self.receiver_ip, self.rmeta_port, self.snap_name))
@@ -143,13 +150,6 @@ class Sender(ReplicationMixin, Process):
         with self._clean_exit_handler(msg):
             self.data_push = self.ctx.socket(zmq.PUSH)
             self.data_push.connect('tcp://%s:%d' % (self.sender_ip, settings.REPLICA_SINK_PORT))
-
-        msg = ('Failed to subscribe to the main scheduler')
-        with self._clean_exit_handler(msg):
-            self.meta_sub = self.ctx.socket(zmq.SUB)
-            self.meta_sub.connect('tcp://%s:%d' % (self.sender_ip, self.sdata_port))
-            self.meta_sub.RCVTIMEO = 600000 #10 minutes
-            self.meta_sub.setsockopt(zmq.SUBSCRIBE, 'meta-%s' % self.snap_id)
 
         #  1. create a new replica trail if it's the very first time
         # or if the last one succeeded
@@ -164,7 +164,7 @@ class Sender(ReplicationMixin, Process):
         #  failed attempt.
         msg = ('Failed to create snapshot: %s. Aborting.' % self.snap_name)
         with self._clean_exit_handler(msg):
-            self.create_snapshot(self.replica.share, self.snap_name)
+            self.create_snapshot(self.replica.share, self.snap_name, self.logger)
 
         #  let the receiver know that following diff is coming
         msg = ('Failed to send initial metadata communication to the '
@@ -173,14 +173,13 @@ class Sender(ReplicationMixin, Process):
         with self._update_trail_and_quit(msg):
             meta_push.send_json(self.meta_begin)
 
-        msg = ('Timeout occured(60 seconds) while waiting for OK '
-               'from the receiver(%s) to start sending data. Aborting.'
-               % self.receiver_ip)
+        msg = ('Timeout occured while waiting for OK '
+               'from the receiver(%s) to start sending data for %s. Aborting.'
+               % (self.receiver_ip, self.snap_id))
         with self._update_trail_and_quit(msg):
             ack = self._process_q()
             if (ack['msg'] == 'snap_exists'):
                 data = {'status': 'succeeded',
-                        'end_ts': datetime.utcnow().replace(tzinfo=utc).strftime(settings.SNAP_TS_FORMAT),
                         'error': 'snapshot already exists on the receiver', }
                 msg = ('Failed to update replica status for snap_name: %s. '
                        'Aborting.' % self.snap_name)
@@ -196,20 +195,20 @@ class Sender(ReplicationMixin, Process):
             prev_snap = ('%s%s/.snapshots/%s/%s' %
                          (settings.MNT_PT, self.replica.pool,
                           self.replica.share, self.rt.snap_name))
-            logger.info('Sending incremental replica between %s -- %s' %
+            self.logger.info('Sending incremental replica between %s -- %s' %
                         (prev_snap, snap_path))
             cmd = [BTRFS, 'send', '-p', prev_snap, snap_path]
         else:
-            logger.info('Sending full replica: %s' % snap_path)
+            self.logger.info('Sending full replica: %s' % snap_path)
 
         try:
-            sp = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE,
-                                  stderr=subprocess.PIPE)
-            fcntl.fcntl(sp.stdout.fileno(), fcntl.F_SETFL, os.O_NONBLOCK)
+            self.sp = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE,
+                                       stderr=subprocess.PIPE)
+            fcntl.fcntl(self.sp.stdout.fileno(), fcntl.F_SETFL, os.O_NONBLOCK)
         except Exception, e:
             msg = ('Failed to start the low level btrfs send '
                    'command(%s). Aborting. Exception: ' % (cmd, e.__str__()))
-            logger.error(msg)
+            self.logger.error(msg)
             with self._update_trail_and_quit(msg):
                 self.data_push.send('%sEND_FAIL' % self.snap_id)
             self._sys_exit(3)
@@ -219,33 +218,38 @@ class Sender(ReplicationMixin, Process):
         check_credit = True
         while alive:
             if (check_credit is True and credit < 1):
-                #This does not seem reliable. @todo: redo.
-                ack = self.meta_sub.recv()
-                ack = json.loads(ack[len('meta-%s' % self.snap_id):])
-                if (ack['msg'] == 'send_more'):
-                    credit = ack['credit']
-                    logger.debug('send process alive for %s. %d KB sent.' %
-                                 (self.snap_id, int(self.kb_sent/1024)))
-                else:
-                    logger.error('unexpected message received: %s' % ack)
+                msg = ('Exception while waiting on receiver to ask for more '
+                       'credit')
+                with self._update_trail_and_quit(msg):
+                    ack = self.meta_sub.recv_string()
+                    ack = json.loads(ack[len('meta-%s' % self.snap_id):])
+                    self.logger.debug('ack received by sender: %s' % ack)
+                    if (ack['msg'] == 'send_more'):
+                        credit = ack['credit']
+                        self.logger.debug('send process alive for %s. %d KB sent.' %
+                                          (self.snap_id, int(self.kb_sent/1024)))
+                    else:
+                        raise Exception('unexpected message received: %s' % ack)
+
             try:
-                if (sp.poll() is not None):
-                    logger.debug('send process finished for %s. rc: %d. '
-                                 'stderr: %s' % (self.snap_id,
-                                                 sp.returncode,
-                                                 sp.stderr.read()))
+                if (self.sp.poll() is not None):
+                    self.logger.debug('send process finished for %s. rc: %d. '
+                                      'stderr: %s' % (self.snap_id,
+                                                      self.sp.returncode,
+                                                      self.sp.stderr.read()))
                     alive = False
-                fs_data = sp.stdout.read()
+                fs_data = self.sp.stdout.read()
             except IOError:
                 continue
             except Exception, e:
                 msg = ('Exception occured while reading low level btrfs '
                        'send data for %s. Aborting.' % self.snap_id)
                 if (alive):
-                    sp.terminate()
+                    self.sp.terminate()
                 with self._update_trail_and_quit(msg):
                     self.data_push.send('%sEND_FAIL' % self.snap_id)
-                    raise e
+                    self.logger.debug(msg)
+                    self._sys_exit(3)
 
             msg = ('Failed to send fsdata to the receiver for %s. Aborting.' %
                    (self.snap_id))
@@ -256,39 +260,38 @@ class Sender(ReplicationMixin, Process):
 
                 if (not alive):
                     check_credit = False
-                    if (sp.returncode != 0):
+                    if (self.sp.returncode != 0):
                         self.data_push.send('%sEND_FAIL' % self.snap_id)
                     else:
                         self.data_push.send('%sEND_SUCCESS' % self.snap_id)
 
             if (os.getppid() != self.ppid):
-                logger.error('Scheduler exited. Sender for %s cannot go on. '
-                             'Aborting.' % self.snap_id)
+                self.logger.error('Scheduler exited. Sender for %s cannot go on. '
+                                  'Aborting.' % self.snap_id)
                 self._sys_exit(3)
 
-        msg = ('Timeout occured(60 seconds) while waiting for final '
+        msg = ('Timeout occured while waiting for final '
                'send confirmation from the receiver(%s) for %s. Aborting.'
                % (self.receiver_ip, self.snap_id))
         with self._update_trail_and_quit(msg):
             ack = self._process_q()
             self.data_push.send('%sACK_SUCCESS' % self.snap_id)
 
-        end_ts = datetime.utcnow().replace(tzinfo=utc).strftime(settings.SNAP_TS_FORMAT)
         data = {'status': 'succeeded',
-                'kb_sent': self.kb_sent / 1024,
-                'end_ts': end_ts, }
+                'kb_sent': self.kb_sent / 1024, }
         if (ack['msg'] == 'receive_error'):
             msg = ('Receiver(%s) returned a processing error for '
                    ' %s. Check it for more information.'
                    % (self.receiver_ip, self.snap_id))
             data['status'] = 'failed'
             data['error'] = msg
-            data['send_failed'] = end_ts
         else:
             share_path = ('%s%s/.snapshots/%s' %
                           (settings.MNT_PT, self.replica.pool,
                            self.replica.share))
-            self._delete_old_snaps(share_path)
+            msg = ('Failed to delete old snapshots')
+            with self._clean_exit_handler(msg):
+                self._delete_old_snaps(share_path)
 
         msg = ('Failed to update final replica status for %s'
                '. Aborting.' % self.snap_id)
